+++ core/src/main/java/io/kestra/core/services/FlowService.java
@@ -1,19 +1,16 @@
 package io.kestra.core.services;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import io.kestra.core.exceptions.FlowProcessingException;
 import io.kestra.core.exceptions.IllegalVariableEvaluationException;
 import io.kestra.core.models.executions.Execution;
 import io.kestra.core.models.flows.*;
 import io.kestra.core.models.flows.check.Check;
-import io.kestra.core.models.tasks.RunnableTask;
 import io.kestra.core.models.topologies.FlowTopology;
 import io.kestra.core.models.triggers.AbstractTrigger;
 import io.kestra.core.models.triggers.TriggerId;
 import io.kestra.core.models.triggers.WorkerTriggerInterface;
 import io.kestra.core.models.validations.ModelValidator;
 import io.kestra.core.models.validations.ValidateConstraintViolation;
-import io.kestra.core.plugins.PluginRegistry;
 import io.kestra.core.queues.QueueException;
 import io.kestra.core.queues.QueueFactoryInterface;
 import io.kestra.core.queues.QueueInterface;
@@ -21,16 +18,14 @@ import io.kestra.core.repositories.FlowRepositoryInterface;
 import io.kestra.core.repositories.FlowTopologyRepositoryInterface;
 import io.kestra.core.runners.RunContext;
 import io.kestra.core.runners.RunContextFactory;
-import io.kestra.core.serializers.JacksonMapper;
+import io.kestra.core.scheduler.TriggerEventQueue;
+import io.kestra.core.scheduler.events.TriggerCreated;
+import io.kestra.core.scheduler.events.TriggerDeleted;
+import io.kestra.core.scheduler.events.TriggerEvent;
+import io.kestra.core.scheduler.events.TriggerUpdated;
 import io.kestra.core.topologies.FlowTopologyService;
 import io.kestra.core.utils.ExecutorsUtils;
 import io.kestra.core.utils.ListUtils;
-import io.kestra.plugin.core.flow.Pause;
-import io.kestra.scheduler.TriggerEventQueue;
-import io.kestra.scheduler.events.TriggerCreated;
-import io.kestra.scheduler.events.TriggerDeleted;
-import io.kestra.scheduler.events.TriggerEvent;
-import io.kestra.scheduler.events.TriggerUpdated;
 import jakarta.annotation.PreDestroy;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
@@ -38,12 +33,8 @@ import jakarta.inject.Provider;
 import jakarta.inject.Singleton;
 import jakarta.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.ClassUtils;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
 import java.time.Instant;
 import java.util.*;
 import java.util.concurrent.ExecutorService;
@@ -51,9 +42,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 
 /**
  * Provides business logic for manipulating flow objects.
@@ -62,19 +51,16 @@ import java.util.stream.StreamSupport;
 @Slf4j
 public class FlowService {
     @Inject
-    private Optional<FlowRepositoryInterface> flowRepository;
+    private FlowRepositoryInterface flowRepository;
 
     @Inject
     private PluginDefaultService pluginDefaultService;
 
     @Inject
-    private PluginRegistry pluginRegistry;
-
-    @Inject
     private ModelValidator modelValidator;
 
     @Inject
-    private Optional<FlowTopologyRepositoryInterface> flowTopologyRepository;
+    private FlowTopologyRepositoryInterface flowTopologyRepository;
 
     @Inject
     private Provider<RunContextFactory> runContextFactory; // Lazy init: avoid circular dependency error.
@@ -88,6 +74,9 @@ public class FlowService {
 
     @Inject
     private TriggerEventQueue triggerEventQueue;
+
+    @Inject
+    private FlowValidationService flowValidationService;
 
     private final ExecutorService executorService;
 
@@ -128,7 +117,7 @@ public class FlowService {
         // properties that might be injecting through default values.
         modelValidator.validate(pluginDefaultService.injectAllDefaults(parsed, false));
 
-        FlowWithSource created = repository().create(flow);
+        FlowWithSource created = flowRepository.create(flow);
 
         // impact downstream consumers: topology, scheduler and flow metastore
         impactDownstreamConsumers(created);
@@ -160,7 +149,7 @@ public class FlowService {
         // properties that might be injecting through default values.
         modelValidator.validate(pluginDefaultService.injectAllDefaults(parsed, false));
 
-        FlowWithSource updated = repository().update(flow, previous);
+        FlowWithSource updated = flowRepository.update(flow, previous);
 
         // impact downstream consumers: topology, scheduler and flow metastore
         impactDownstreamConsumers(updated);
@@ -172,11 +161,7 @@ public class FlowService {
      * Delete a flow.
      */
     public FlowWithSource delete(FlowWithSource flow) throws QueueException {
-        if (flowRepository.isEmpty()) {
-            throw noRepositoryException();
-        }
-
-        FlowWithSource deleted = flowRepository.get().delete(flow);
+        FlowWithSource deleted = flowRepository.delete(flow);
 
         // impact downstream consumers: topology, scheduler and flow metastore
         impactDownstreamConsumers(deleted);
@@ -196,14 +181,14 @@ public class FlowService {
     }
 
     private void updateTopology(FlowWithSource flow) {
-        flowTopologyRepository.get().save(
+        flowTopologyRepository.save(
             flow,
             (flow.isDeleted() ?
                 Stream.<FlowTopology>empty() :
                 flowTopologyService
                     .topology(
                         flow,
-                        flowRepository.get().findAllWithSource(flow.getTenantId())
+                        flowRepository.findAllWithSource(flow.getTenantId())
                     )
             )
                 .distinct()
@@ -212,7 +197,7 @@ public class FlowService {
     }
 
     private void recomputeTriggers(FlowWithSource flow) {
-        var previous = flow.getRevision() <= 1 ? null : flowRepository.get().findById(flow.getTenantId(), flow.getNamespace(), flow.getId(), Optional.of(flow.getRevision() - 1)).orElse(null);
+        var previous = flow.getRevision() <= 1 ? null : flowRepository.findById(flow.getTenantId(), flow.getNamespace(), flow.getId(), Optional.of(flow.getRevision() - 1)).orElse(null);
 
         if (flow.isDeleted() || previous != null) {
             List<AbstractTrigger> triggersDeleted = flow.isDeleted() ?
@@ -220,7 +205,7 @@ public class FlowService {
                 FlowService.findRemovedTrigger(flow, previous);
 
             triggersDeleted.forEach(trigger ->
-                sendTriggerEvent(new TriggerDeleted(TriggerId.of(flow, trigger), Instant.now()))
+                sendTriggerEvent(new TriggerDeleted(TriggerId.of(flow, trigger)))
             );
         }
 
@@ -229,13 +214,13 @@ public class FlowService {
                 .stream()
                 .filter(trigger -> trigger instanceof WorkerTriggerInterface)
                 .forEach(trigger ->
-                    sendTriggerEvent(new TriggerUpdated(TriggerId.of(flow, trigger), flow.getRevision(), Instant.now()))
+                    sendTriggerEvent(new TriggerUpdated(TriggerId.of(flow, trigger), flow.getRevision()))
                 );
             FlowService.findNewTrigger(flow, previous)
                 .stream()
                 .filter(trigger -> trigger instanceof WorkerTriggerInterface)
                 .forEach(trigger ->
-                    sendTriggerEvent(new TriggerUpdated(TriggerId.of(flow, trigger), flow.getRevision(), Instant.now()))
+                    sendTriggerEvent(new TriggerUpdated(TriggerId.of(flow, trigger), flow.getRevision()))
                 );
             return;
         }
@@ -245,18 +230,13 @@ public class FlowService {
                 .stream()
                 .filter(trigger -> trigger instanceof WorkerTriggerInterface)
                 .forEach(trigger ->
-                    sendTriggerEvent(new TriggerCreated(TriggerId.of(flow, trigger), Instant.now(), flow.getRevision()))
+                    sendTriggerEvent(new TriggerCreated(TriggerId.of(flow, trigger), flow.getRevision()))
                 );
         }
     }
 
     private void sendTriggerEvent(TriggerEvent event) {
         this.triggerEventQueue.send(event);
-    }
-
-    private FlowRepositoryInterface repository() {
-        return flowRepository
-            .orElseThrow(() -> new IllegalStateException("Cannot perform operation on flow. Cause: No FlowRepository"));
     }
 
     /**
@@ -304,9 +284,9 @@ public class FlowService {
     }
 
     /**
-     * Validates the given flow source.
+     * Validates the given flow source. The YAML source can contain one or many objects.
      * <p>
-     * the YAML source can contain one or many objects.
+     * Individual validation methods are provided inside the {@link FlowValidationService}.
      *
      * @param tenantId  The tenant identifier.
      * @param flows     The YAML source.
@@ -324,14 +304,14 @@ public class FlowService {
                     FlowWithSource flow = pluginDefaultService.parseFlowWithVersionDefaults(tenantId, source, true);
                     Integer sentRevision = flow.getRevision();
                     if (sentRevision != null) {
-                        Integer lastRevision = Optional.ofNullable(repository().lastRevision(tenantId, flow.getNamespace(), flow.getId()))
+                        Integer lastRevision = Optional.ofNullable(flowRepository.lastRevision(tenantId, flow.getNamespace(), flow.getId()))
                             .orElse(0);
                         validateConstraintViolationBuilder.outdated(!sentRevision.equals(lastRevision + 1));
                     }
 
-                    validateConstraintViolationBuilder.deprecationPaths(deprecationPaths(flow));
-                    validateConstraintViolationBuilder.warnings(warnings(flow, tenantId));
-                    validateConstraintViolationBuilder.infos(relocations(source).stream().map(relocation -> relocation.from() + " is replaced by " + relocation.to()).toList());
+                    validateConstraintViolationBuilder.deprecationPaths(flowValidationService.deprecationPaths(flow));
+                    validateConstraintViolationBuilder.warnings(flowValidationService.warnings(flow, tenantId));
+                    validateConstraintViolationBuilder.infos(flowValidationService.relocations(source).stream().map(relocation -> relocation.from() + " is replaced by " + relocation.to()).toList());
                     validateConstraintViolationBuilder.flow(flow.getId());
                     validateConstraintViolationBuilder.namespace(flow.getNamespace());
 
@@ -368,7 +348,7 @@ public class FlowService {
 
         final GenericFlow flow = GenericFlow.fromYaml(tenantId, source);
 
-        Optional<FlowWithSource> maybeExisting = repository().findByIdWithSource(
+        Optional<FlowWithSource> maybeExisting = flowRepository.findByIdWithSource(
             flow.getTenantId(),
             flow.getNamespace(),
             flow.getId(),
@@ -389,202 +369,25 @@ public class FlowService {
                 .orElseGet(() -> FlowWithSource.of(flowToImport, source).toBuilder().tenantId(tenantId).revision(1).build());
         } else {
             return maybeExisting
-                .map(previous -> repository().update(flow, previous))
-                .orElseGet(() -> repository().create(flow));
+                .map(previous -> flowRepository.update(flow, previous))
+                .orElseGet(() -> flowRepository.create(flow));
         }
     }
 
     public List<FlowWithSource> findByNamespaceWithSource(String tenantId, String namespace) {
-        if (flowRepository.isEmpty()) {
-            throw noRepositoryException();
-        }
-
-        return flowRepository.get().findByNamespaceWithSource(tenantId, namespace);
+        return flowRepository.findByNamespaceWithSource(tenantId, namespace);
     }
 
     public List<Flow> findAll(String tenantId) {
-        if (flowRepository.isEmpty()) {
-            throw noRepositoryException();
-        }
-
-        return flowRepository.get().findAll(tenantId);
+        return flowRepository.findAll(tenantId);
     }
 
     public List<Flow> findByNamespace(String tenantId, String namespace) {
-        if (flowRepository.isEmpty()) {
-            throw noRepositoryException();
-        }
-
-        return flowRepository.get().findByNamespace(tenantId, namespace);
+        return flowRepository.findByNamespace(tenantId, namespace);
     }
 
     public Optional<Flow> findById(String tenantId, String namespace, String flowId) {
-        if (flowRepository.isEmpty()) {
-            throw noRepositoryException();
-        }
-
-        return flowRepository.get().findById(tenantId, namespace, flowId);
-    }
-
-    public List<String> deprecationPaths(Flow flow) {
-        return deprecationTraversal("", flow).toList();
-    }
-
-
-    public List<String> warnings(Flow flow, String tenantId) {
-        if (flow == null) {
-            return Collections.emptyList();
-        }
-
-        List<String> warnings = new ArrayList<>(checkValidSubflows(flow, tenantId));
-
-        List<io.kestra.plugin.core.trigger.Flow> flowTriggers = ListUtils.emptyOnNull(flow.getTriggers()).stream()
-            .filter(io.kestra.plugin.core.trigger.Flow.class::isInstance)
-            .map(io.kestra.plugin.core.trigger.Flow.class::cast)
-            .toList();
-        flowTriggers.forEach(flowTrigger -> {
-            if (ListUtils.emptyOnNull(flowTrigger.getConditions()).isEmpty() && flowTrigger.getPreconditions() == null) {
-                warnings.add("This flow will be triggered for EVERY execution of EVERY flow on your instance. We recommend adding the preconditions property to the Flow trigger '" + flowTrigger.getId() + "'.");
-            }
-        });
-
-        // add warning for runnable properties (timeout, workerGroup, taskCache) when used not in a runnable
-        flow.allTasksWithChilds().forEach(task -> {
-            if (!(task instanceof RunnableTask<?>)) {
-                if (task.getTimeout() != null && !(task instanceof Pause)) {
-                    warnings.add("The task '" + task.getId() + "' cannot use the 'timeout' property as it's only relevant for runnable tasks.");
-                }
-                if (task.getTaskCache() != null) {
-                    warnings.add("The task '" + task.getId() + "' cannot use the 'taskCache' property as it's only relevant for runnable tasks.");
-                }
-                if (task.getWorkerGroup() != null) {
-                    warnings.add("The task '" + task.getId() + "' cannot use the 'workerGroup' property as it's only relevant for runnable tasks.");
-                }
-            }
-        });
-
-        return warnings;
-    }
-
-    public List<Relocation> relocations(String flowSource) {
-        try {
-            Map<String, Class<?>> aliases = pluginRegistry.plugins().stream()
-                .flatMap(plugin -> plugin.getAliases().values().stream())
-                .collect(Collectors.toMap(
-                    Map.Entry::getKey,
-                    Map.Entry::getValue,
-                    (existing, duplicate) -> existing
-                ));
-            Map<String, Object> stringObjectMap = JacksonMapper.ofYaml().readValue(flowSource, JacksonMapper.MAP_TYPE_REFERENCE);
-            return relocations(aliases, stringObjectMap);
-        } catch (JsonProcessingException e) {
-            // silent failure (we don't compromise the app / response for warnings)
-            return Collections.emptyList();
-        }
-    }
-
-    // check if subflow is present in given namespace
-    public List<String> checkValidSubflows(Flow flow, String tenantId) {
-        List<io.kestra.plugin.core.flow.Subflow> subFlows = ListUtils.emptyOnNull(flow.getTasks()).stream()
-            .filter(io.kestra.plugin.core.flow.Subflow.class::isInstance)
-            .map(io.kestra.plugin.core.flow.Subflow.class::cast)
-            .toList();
-
-        List<String> violations = new ArrayList<>();
-
-        subFlows.forEach(subflow -> {
-            String regex = ".*\\{\\{.+}}.*"; // regex to check if string contains pebble
-            String subflowId = subflow.getFlowId();
-            String namespace = subflow.getNamespace();
-            if ((subflowId != null && subflowId.matches(regex)) || (namespace != null && namespace.matches(regex))) {
-                return;
-            }
-            if (subflowId == null || namespace == null) {
-                // those fields are mandatory so the mandatory validation will apply
-                return;
-            }
-            Optional<Flow> optional = findById(tenantId, subflow.getNamespace(), subflow.getFlowId());
-
-            if (optional.isEmpty()) {
-                violations.add("The subflow '" + subflow.getFlowId() + "' not found in namespace '" + subflow.getNamespace() + "'.");
-            } else if (optional.get().isDisabled()) {
-                violations.add("The subflow '" + subflow.getFlowId() + "' is disabled in namespace '" + subflow.getNamespace() + "'.");
-            }
-        });
-
-        return violations;
-    }
-
-    public record Relocation(String from, String to) {}
-
-    @SuppressWarnings("unchecked")
-    private List<Relocation> relocations(Map<String, Class<?>> aliases, Map<String, Object> stringObjectMap) {
-        List<Relocation> relocations = new ArrayList<>();
-        for (Map.Entry<String, Object> entry : stringObjectMap.entrySet()) {
-            if (entry.getValue() instanceof String value && aliases.containsKey(value)) {
-                relocations.add(new Relocation(value, aliases.get(value).getName()));
-            }
-
-            if (entry.getValue() instanceof Map<?, ?> value) {
-                relocations.addAll(relocations(aliases, (Map<String, Object>) value));
-            }
-
-            if (entry.getValue() instanceof List<?> value) {
-                List<Relocation> listAliases = value.stream().flatMap(item -> {
-                    if (item instanceof Map<?, ?> map) {
-                        return relocations(aliases, (Map<String, Object>) map).stream();
-                    }
-                    return Stream.empty();
-                }).toList();
-                relocations.addAll(listAliases);
-            }
-        }
-
-        return relocations;
-    }
-
-
-    private Stream<String> deprecationTraversal(String prefix, Object object) {
-        if (object == null || ClassUtils.isPrimitiveOrWrapper(object.getClass()) || String.class.equals(object.getClass())) {
-            return Stream.empty();
-        }
-
-        return Stream.concat(
-            object.getClass().isAnnotationPresent(Deprecated.class) ? Stream.of(prefix) : Stream.empty(),
-            allGetters(object.getClass())
-                .flatMap(method -> {
-                    try {
-                        Object fieldValue = method.invoke(object);
-
-                        if (fieldValue instanceof Iterable<?> iterableValue) {
-                            fieldValue = StreamSupport.stream(iterableValue.spliterator(), false).toArray(Object[]::new);
-                        }
-
-                        String fieldName = method.getName().substring(3, 4).toLowerCase() + method.getName().substring(4);
-                        Stream<String> additionalDeprecationPaths = Stream.empty();
-                        if (fieldValue instanceof Object[] arrayValue) {
-                            additionalDeprecationPaths = IntStream.range(0, arrayValue.length).boxed().flatMap(i -> deprecationTraversal(fieldName + "[%d]".formatted(i), arrayValue[i]));
-                        }
-
-                        return Stream.concat(
-                            method.isAnnotationPresent(Deprecated.class) && fieldValue != null ? Stream.of(prefix.isEmpty() ? fieldName : prefix + "." + fieldName) : Stream.empty(),
-                            additionalDeprecationPaths
-                        );
-                    } catch (IllegalAccessException | InvocationTargetException e) {
-                        // silent failure (we don't compromise the app / response for warnings)
-                    }
-
-                    return Stream.empty();
-                })
-        );
-    }
-
-    private Stream<Method> allGetters(Class<?> clazz) {
-        return Arrays.stream(clazz.getMethods())
-            .filter(m -> !m.getDeclaringClass().equals(Object.class))
-            .filter(method -> method.getName().startsWith("get") && method.getName().length() > 3 && method.getParameterCount() == 0)
-            .filter(method -> !method.getReturnType().equals(Void.TYPE))
-            .filter(method -> !Modifier.isStatic(method.getModifiers()));
+        return flowRepository.findById(tenantId, namespace, flowId);
     }
 
     public boolean removeUnwanted(Flow f, Execution execution) {
@@ -639,11 +442,7 @@ public class FlowService {
 
     // Used in Git plugin
     public List<Flow> findByNamespacePrefix(String tenantId, String namespacePrefix) {
-        if (flowRepository.isEmpty()) {
-            throw noRepositoryException();
-        }
-
-        return flowRepository.get().findByNamespacePrefix(tenantId, namespacePrefix);
+        return flowRepository.findByNamespacePrefix(tenantId, namespacePrefix);
     }
 
     /**
@@ -659,11 +458,7 @@ public class FlowService {
      * @throws IllegalStateException  if the requested flow is not executable.
      */
     public Flow getFlowIfExecutableOrThrow(final String tenant, final String namespace, final String id, final Optional<Integer> revision) {
-        if (flowRepository.isEmpty()) {
-            throw noRepositoryException();
-        }
-
-        Optional<Flow> optional = flowRepository.get().findByIdWithoutAcl(tenant, namespace, id, revision);
+        Optional<Flow> optional = flowRepository.findByIdWithoutAcl(tenant, namespace, id, revision);
         if (optional.isEmpty()) {
             throw new NoSuchElementException("Requested Flow is not found.");
         }
@@ -680,19 +475,11 @@ public class FlowService {
     }
 
     public Stream<FlowTopology> findDependencies(final String tenant, final String namespace, final String id, boolean destinationOnly, boolean expandAll) {
-        if (flowTopologyRepository.isEmpty()) {
-            throw noRepositoryException();
-        }
-
-        return expandAll ? recursiveFlowTopology(new ArrayList<>(), tenant, namespace, id, destinationOnly) : flowTopologyRepository.get().findByFlow(tenant, namespace, id, destinationOnly).stream();
+        return expandAll ? recursiveFlowTopology(new ArrayList<>(), tenant, namespace, id, destinationOnly) : flowTopologyRepository.findByFlow(tenant, namespace, id, destinationOnly).stream();
     }
 
     private Stream<FlowTopology> recursiveFlowTopology(List<String> visitedTopologies, String tenantId, String namespace, String id, boolean destinationOnly) {
-        if (flowTopologyRepository.isEmpty()) {
-            throw noRepositoryException();
-        }
-
-        var flowTopologies = flowTopologyRepository.get().findByFlow(tenantId, namespace, id, destinationOnly);
+        var flowTopologies = flowTopologyRepository.findByFlow(tenantId, namespace, id, destinationOnly);
 
         var visitedNodes = new ArrayList<String>();
         visitedNodes.add(id);
