+++ webserver/src/test/java/io/kestra/webserver/controllers/api/ExecutionControllerTest.java
@@ -540,7 +540,7 @@ class ExecutionControllerTest {
             .build();
 
         client.toBlocking().retrieve(
-            HttpRequest.POST("/api/v1/main/flows", create),
+            HttpRequest.POST("/api/v1/main/flows", create.sourceOrGenerateIfNull()).contentType(MediaType.APPLICATION_YAML_TYPE),
             Flow.class
         );
